Split order summary row model file
